% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{basis}
\alias{basis}
\title{Bivariate Spline Basis Function}
\usage{
basis(V, Tr, d = 5, r = 1, Z, Hmtx = TRUE, Kmtx = TRUE, QR = TRUE, TA = TRUE)
}
\arguments{
\item{V}{The \code{N} by two matrix of vertices of a triangulation, where \code{N} is the number of vertices. Each row is the coordinates for a vertex.
\cr}

\item{Tr}{The triangulation matrix of dimention \code{nT} by three, where \code{nT} is the number of triangles in the triangulation. Each row is the indices of vertices in \code{V}.
\cr}

\item{d}{The degree of piecewise polynomials -- default is 5, and usually \code{d} is greater than one. -1 represents piecewise constant.
\cr}

\item{r}{The smoothness parameter -- default is 1, and 0 \eqn{\le} \code{r} \eqn{<} \code{d}.
\cr}

\item{Z}{The cooridinates of dimension \code{n} by two. Each row is the coordinates of a point.
\cr}

\item{Hmtx}{The indicator of whether the smoothness matrix \code{H} need to be generated -- default is \code{TRUE}.
\cr}

\item{Kmtx}{The indicator of whether the energy matrix \code{K} need to be generated -- default is \code{TRUE}.
\cr}

\item{QR}{The indicator of whether a QR decomposition need to be performed on the smoothness matrix -- default is \code{TRUE}.
\cr}

\item{TA}{The indicator of whether the area of the triangles need to be calculated -- default is \code{TRUE}.
\cr}
}
\value{
A list of vectors and matrice, including:
\item{B}{The spline basis function of dimension \code{n} by \code{nT}*\code{{(d+1)(d+2)/2}}, where \code{n} is the number of observationed points, \code{nT} is the number of triangles in the given triangulation, and \code{d} is the degree of the spline. If some points do not fall in the triangulation, the generation of the spline basis will not take those points into consideration.}
\item{Ind.inside}{A vector contains the indexes of all the points which are inside the triangulation.}
\item{H}{The smoothness matrix.}
\item{Q2}{The Q2 matrix after QR decomposition of the smoothness matrix \code{H}.}
\item{K}{The thin-plate energy function.}
\item{tria.all}{The area of each triangle within the given triangulation.}
}
\description{
This function generates the basis for bivariate spline over triangulation.
}
\details{
This R program is modified based on the Matlab program written by Ming-Jun Lai from the University of Georgia and Li Wang from the Iowa State University.
}
\examples{
# example 1
xx = c(-0.25, 0.75, 0.25, 1.25)
yy = c(-0.25, 0.25, 0.75,1 .25)
Z = cbind(xx, yy)
d = 4; r = 1;
V0 = rbind(c(0, 0), c(1, 0), c(1, 1), c(0, 1))
Tr0 = rbind(c(1, 2, 3), c(1, 3, 4))
basis(V0, Tr0, d, r, Z)
}
