% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.BPST.R
\name{data.BPST}
\alias{data.BPST}
\title{Generate testing dataset for bivariate spline smoothing.}
\usage{
data.BPST(Z, V, Tr, func = 1, sigma = 0.1, iter = 2019)
}
\arguments{
\item{Z}{The cooridinates of dimension \code{n} by two. Each row is the coordinates of a point.
\cr}

\item{V}{The \code{N} by two matrix of vertices of a triangulation, where \code{N} is the number of vertices. Each row is the coordinates for a vertex.
\cr}

\item{Tr}{The triangulation matrix of dimention \code{nT} by three, where \code{nT} is the number of triangles in the triangulation. Each row is the indices of vertices in \code{V}.
\cr}

\item{func}{The choice of test function -- default is 1. Possible choices include 1,2,3,4,5,6,7,8.
\cr}

\item{sigma}{The standard deviation of the white noise --  default is 0.1.
\cr}
}
\value{
A list of vectors and matrice, including:
\item{Y}{The response variable.}
\item{mu}{The mean function.}
\item{Z}{The coordinates.}
\item{ind}{A vector contains the indicators whether the point is inside the given triangulation.}
}
\description{
This function generate the testing dataset for bivariate spline smoothing.
}
\details{
This R program is modified based on the Matlab program written by Ming-Jun Lai from the University of Georgia and Li Wang from the Iowa State University.
}
