% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inVT.R
\name{inVT}
\alias{inVT}
\title{Decide whether a point is inside of a given triangulation.}
\usage{
inVT(V0, Tr0, xx, yy)
}
\arguments{
\item{V0}{The \code{N} by two matrix of vertices of a triangulation, where \code{N} is the number of vertices. Each row is the coordinates for a vertex.
\cr}

\item{Tr0}{The triangulation matrix of dimention \code{nT} by three, where \code{nT} is the number of triangles in the triangulation. Each row is the indices of vertices in \code{V}.
\cr}

\item{xx}{The x-cooridinate of points of dimension \code{n} by one. 
\cr}

\item{yy}{The y-cooridinate of points of dimension \code{n} by one.
\cr}
}
\value{
A list of vectors, including:
\item{ind}{A vector of dimension \code{n} by one matrix that lists whether the points are inside of a given triangulation. 0 -- represents outside the triangulation, while 1 -- represents inside the triangulation.}
\item{ind.inside}{A vector contains the indexes of all the points which are inside the triangulation.}
}
\description{
This function is used to decided whether a point is inside of a given triangulation.
}
\details{
This R program is modified based on the Matlab program written by Ming-Jun Lai from the University of Georgia and Li Wang from the Iowa State University.
}
\examples{
xx=c(-0.25,0.75,0.25,1.25)
yy=c(-0.25,0.25,0.75,1.25)
V0=rbind(c(0,0),c(1,0),c(1,1),c(0,1))
Tr0=rbind(c(1,2,3),c(1,3,4))
inVT(V0,Tr0,xx,yy)

}
