% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BPST.R
\name{plot.BPST}
\alias{plot.BPST}
\title{Produces the contour plot for the estimated surface of a fitted "BPST" object.}
\usage{
\method{plot}{BPST}(mfit, Zgrid = NULL)
}
\arguments{
\item{mfit}{Fitted ``BPST" object.
\cr}

\item{Zgrid}{The grid points used to construct the contour plot.
\cr}
}
\value{
None
}
\description{
This function produces the contour map for the estimated surface of a fitted "BPST" object.
}
\details{
This R program is modified based on the Matlab program written by Ming-Jun Lai from the University of Georgia and Li Wang from the Iowa State University.
}
