% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tr1.R
\docType{data}
\name{Tr1}
\alias{Tr1}
\title{Example of triangulation of a rectangular domain.}
\format{\code{Tr1} is a matrix of a triangulation over a rectangular domain.}
\usage{
data('Tr1')
}
\description{
A matrix of a triangulation over a rectangular domain.
}
\references{
This example is generated from R package ``Triangulation".
}
\keyword{datasets}
