% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tr2.R
\docType{data}
\name{Tr2}
\alias{Tr2}
\title{Example of triangulation of an irregular domain.}
\format{\code{Tr2} is a matrix of a triangulation over an irregular domain.}
\usage{
data('Tr2')
}
\description{
A matrix of a triangulation over an irregular domain.
}
\references{
This example is generated from R package ``Triangulation".
}
\keyword{datasets}
