% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1.R
\docType{data}
\name{V1}
\alias{V1}
\title{Example of vertices of a rectangular domain.}
\format{
\code{V1} is a matrix of coordinates of the vertices over a rectangular domain.
}
\usage{
data('V1')
}
\description{
A matrix of coordinates of the vertices over a rectangular domain.
}
\references{
This example is generated from R package ``Triangulation".
}
\keyword{datasets}
