% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2.R
\docType{data}
\name{V2}
\alias{V2}
\title{Example of vertices of an irregular domain.}
\format{
\code{V2} is a matrix of coordinates of the vertices over an irregular domain.
}
\usage{
data('V2')
}
\description{
A matrix of coordinates of the vertices over an irregular domain.
}
\references{
This example is generated from R package ``Triangulation".
}
\keyword{datasets}
