% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.BPST.R
\name{predict.BPST}
\alias{predict.BPST}
\title{Make predictions from a fitted BPST object.}
\usage{
\method{predict}{BPST}(mfit, Zpred = NULL)
}
\arguments{
\item{mfit}{Fitted ``BPST" object.
\cr}

\item{Zpred}{The cooridinates for prediction -- default is the observed coordinates, \code{Z}.
\cr}
}
\value{
A vector of predicted values is returned.
}
\description{
This function is used to make predictions of a fitted BPST object.
}
\details{
This R program is modified based on the Matlab program written by Ming-Jun Lai from the University of Georgia ang Li Wang from the Iowa State University.
}
